feat(eventrouter): route via attribute-based matcher (translate-on-read)
Cut Router.Route over to the eventrouter2 attribute-based matcher. Legacy
rules stay in the stored model.RoutingRule shape; they are translated to the
new shape at match time via SchemaRegistry.TranslateRule and matched with
RoutingRule.Match. No proto/model/store/webui changes.

The event is converted once per Route call with InputEvent.toV2 (a field copy
carrying Attrs), and the first-match-per-org loop tests each rule through the
matchesV2 helper. The default {Prefix:"xagent:"} fallback flows through the
same translate path.

Matching fans out over an eventrouter2.SchemaRegistry. Router gains an
injectable Registry field that nil-defaults to eventrouter2.DefaultSchemaRegistry
(populated by the producer packages' init), so production construction sites
are unchanged while tests inject an isolated registry. The producer schema
helpers are exported (RegisterSchemas) so tests build that registry from the
real schemas instead of hand-copied defs.

The v1 matcher (MatchRule/matchMention/matchAssignee) and its test are removed.
Mention/assignee matching is now case-sensitive (the v1 case-insensitivity was
a bug); tests assert the corrected behavior directly.

Because the producer packages import eventrouter, the Route tests live in an
external eventrouter_test package that imports them to build the registry.

Part of #1225
