+++ internal/eventrouter/rule_test.go
@@ -3,302 +3,8 @@ package eventrouter
 import (
 	"testing"
 
-	"github.com/icholy/xagent/internal/model"
 	"gotest.tools/v3/assert"
 )
-
-func TestMatchRule(t *testing.T) {
-	tests := []struct {
-		name  string
-		rule  model.RoutingRule
-		event InputEvent
-		want  bool
-	}{
-		{
-			name:  "empty rule matches everything",
-			rule:  model.RoutingRule{},
-			event: InputEvent{Source: "github", Type: "issue_comment", Data: "hello"},
-			want:  true,
-		},
-		{
-			name:  "source match",
-			rule:  model.RoutingRule{Source: "github"},
-			event: InputEvent{Source: "github"},
-			want:  true,
-		},
-		{
-			name:  "source mismatch",
-			rule:  model.RoutingRule{Source: "github"},
-			event: InputEvent{Source: "atlassian"},
-			want:  false,
-		},
-		{
-			name:  "type match",
-			rule:  model.RoutingRule{Type: "issue_comment"},
-			event: InputEvent{Source: "github", Type: "issue_comment"},
-			want:  true,
-		},
-		{
-			name:  "type mismatch",
-			rule:  model.RoutingRule{Type: "issue_comment"},
-			event: InputEvent{Source: "github", Type: "pull_request_review"},
-			want:  false,
-		},
-		{
-			name:  "prefix match",
-			rule:  model.RoutingRule{Prefix: "xagent:"},
-			event: InputEvent{Data: "xagent: fix the tests"},
-			want:  true,
-		},
-		{
-			name:  "prefix mismatch",
-			rule:  model.RoutingRule{Prefix: "xagent:"},
-			event: InputEvent{Data: "just a comment"},
-			want:  false,
-		},
-		{
-			name:  "github mention match",
-			rule:  model.RoutingRule{Mention: "botuser"},
-			event: InputEvent{Source: "github", Data: "hey @botuser fix this"},
-			want:  true,
-		},
-		{
-			name:  "github mention case insensitive",
-			rule:  model.RoutingRule{Mention: "BotUser"},
-			event: InputEvent{Source: "github", Data: "hey @botuser fix this"},
-			want:  true,
-		},
-		{
-			name:  "github mention no substring match",
-			rule:  model.RoutingRule{Mention: "bot"},
-			event: InputEvent{Source: "github", Data: "hey @botuser fix this"},
-			want:  false,
-		},
-		{
-			name:  "github mention at start",
-			rule:  model.RoutingRule{Mention: "botuser"},
-			event: InputEvent{Source: "github", Data: "@botuser fix this"},
-			want:  true,
-		},
-		{
-			name:  "github mention mismatch",
-			rule:  model.RoutingRule{Mention: "botuser"},
-			event: InputEvent{Source: "github", Data: "no mention here"},
-			want:  false,
-		},
-		{
-			name:  "atlassian mention match",
-			rule:  model.RoutingRule{Mention: "abc123"},
-			event: InputEvent{Source: "atlassian", Data: "hey [~accountid:abc123] fix this"},
-			want:  true,
-		},
-		{
-			name:  "atlassian mention mismatch",
-			rule:  model.RoutingRule{Mention: "abc123"},
-			event: InputEvent{Source: "atlassian", Data: "no mention here"},
-			want:  false,
-		},
-		{
-			name:  "unknown source mention never matches",
-			rule:  model.RoutingRule{Mention: "user"},
-			event: InputEvent{Source: "unknown", Data: "@user hello"},
-			want:  false,
-		},
-		{
-			name:  "all fields must match",
-			rule:  model.RoutingRule{Source: "github", Type: "issue_comment", Prefix: "xagent:"},
-			event: InputEvent{Source: "github", Type: "issue_comment", Data: "xagent: do it"},
-			want:  true,
-		},
-		{
-			name:  "all fields source mismatch",
-			rule:  model.RoutingRule{Source: "github", Type: "issue_comment", Prefix: "xagent:"},
-			event: InputEvent{Source: "atlassian", Type: "issue_comment", Data: "xagent: do it"},
-			want:  false,
-		},
-		{
-			name:  "all fields prefix mismatch",
-			rule:  model.RoutingRule{Source: "github", Type: "issue_comment", Prefix: "xagent:"},
-			event: InputEvent{Source: "github", Type: "issue_comment", Data: "just a comment"},
-			want:  false,
-		},
-		{
-			name:  "github assignee match",
-			rule:  model.RoutingRule{Assignee: "icholy-bot"},
-			event: InputEvent{Source: "github", Type: "pull_request_assigned", Assignee: "icholy-bot"},
-			want:  true,
-		},
-		{
-			name:  "github assignee case insensitive",
-			rule:  model.RoutingRule{Assignee: "icholy-bot"},
-			event: InputEvent{Source: "github", Type: "pull_request_assigned", Assignee: "Icholy-Bot"},
-			want:  true,
-		},
-		{
-			name:  "github assignee mismatch",
-			rule:  model.RoutingRule{Assignee: "icholy-bot"},
-			event: InputEvent{Source: "github", Type: "pull_request_assigned", Assignee: "octocat"},
-			want:  false,
-		},
-		{
-			name:  "github assignee empty on non-assignment event",
-			rule:  model.RoutingRule{Assignee: "icholy-bot"},
-			event: InputEvent{Source: "github", Type: "issue_comment", Data: "hi", Assignee: ""},
-			want:  false,
-		},
-		{
-			name:  "atlassian assignee never matches v1",
-			rule:  model.RoutingRule{Assignee: "5b10ac8d82e05b22cc7d4ef5"},
-			event: InputEvent{Source: "atlassian", Assignee: "5b10ac8d82e05b22cc7d4ef5"},
-			want:  false,
-		},
-		{
-			name:  "unknown source assignee never matches",
-			rule:  model.RoutingRule{Assignee: "user"},
-			event: InputEvent{Source: "unknown", Assignee: "user"},
-			want:  false,
-		},
-		{
-			name:  "empty url prefix matches any url",
-			rule:  model.RoutingRule{},
-			event: InputEvent{URL: "https://github.com/icholy/xagent/pull/1"},
-			want:  true,
-		},
-		{
-			name:  "url prefix match",
-			rule:  model.RoutingRule{URLPrefix: "https://github.com/icholy/xagent/"},
-			event: InputEvent{URL: "https://github.com/icholy/xagent/pull/1"},
-			want:  true,
-		},
-		{
-			name:  "url prefix mismatch",
-			rule:  model.RoutingRule{URLPrefix: "https://github.com/icholy/xagent/"},
-			event: InputEvent{URL: "https://github.com/other/repo/pull/1"},
-			want:  false,
-		},
-		{
-			name:  "url prefix mismatch on empty url",
-			rule:  model.RoutingRule{URLPrefix: "https://github.com/icholy/xagent/"},
-			event: InputEvent{URL: ""},
-			want:  false,
-		},
-		{
-			name: "url prefix is independent of body prefix",
-			rule: model.RoutingRule{URLPrefix: "https://github.com/icholy/xagent/"},
-			event: InputEvent{
-				URL:  "https://github.com/icholy/xagent/issues/1",
-				Data: "no xagent: prefix here",
-			},
-			want: true,
-		},
-		{
-			name: "body prefix is independent of url",
-			rule: model.RoutingRule{Prefix: "xagent:"},
-			event: InputEvent{
-				URL:  "https://example.com/whatever",
-				Data: "xagent: do the thing",
-			},
-			want: true,
-		},
-		{
-			name:  "value match",
-			rule:  model.RoutingRule{Value: "xagent"},
-			event: InputEvent{Values: []string{"xagent", "urgent"}},
-			want:  true,
-		},
-		{
-			name:  "value no match",
-			rule:  model.RoutingRule{Value: "xagent"},
-			event: InputEvent{Values: []string{"bug", "urgent"}},
-			want:  false,
-		},
-		{
-			name:  "empty value is a wildcard",
-			rule:  model.RoutingRule{},
-			event: InputEvent{Values: []string{"xagent"}},
-			want:  true,
-		},
-		{
-			name:  "value with empty event values no match",
-			rule:  model.RoutingRule{Value: "xagent"},
-			event: InputEvent{Values: nil},
-			want:  false,
-		},
-		{
-			name:  "value requires exact membership not substring",
-			rule:  model.RoutingRule{Value: "agent"},
-			event: InputEvent{Values: []string{"xagent"}},
-			want:  false,
-		},
-		{
-			name: "combined source type value all match",
-			rule: model.RoutingRule{Source: "atlassian", Type: "label_added", Value: "xagent"},
-			event: InputEvent{
-				Source: "atlassian",
-				Type:   "label_added",
-				Values: []string{"xagent"},
-			},
-			want: true,
-		},
-		{
-			name: "combined value mismatch",
-			rule: model.RoutingRule{Source: "atlassian", Type: "label_added", Value: "xagent"},
-			event: InputEvent{
-				Source: "atlassian",
-				Type:   "label_added",
-				Values: []string{"bug"},
-			},
-			want: false,
-		},
-		{
-			name: "combined source type assignee all match",
-			rule: model.RoutingRule{
-				Source:   "github",
-				Type:     "pull_request_assigned",
-				Assignee: "icholy-bot",
-			},
-			event: InputEvent{
-				Source:   "github",
-				Type:     "pull_request_assigned",
-				Assignee: "icholy-bot",
-			},
-			want: true,
-		},
-		{
-			name: "combined gate type mismatch",
-			rule: model.RoutingRule{
-				Source:   "github",
-				Type:     "pull_request_assigned",
-				Assignee: "icholy-bot",
-			},
-			event: InputEvent{
-				Source:   "github",
-				Type:     "issue_assigned",
-				Assignee: "icholy-bot",
-			},
-			want: false,
-		},
-		{
-			name: "combined gate assignee mismatch",
-			rule: model.RoutingRule{
-				Source:   "github",
-				Type:     "pull_request_assigned",
-				Assignee: "icholy-bot",
-			},
-			event: InputEvent{
-				Source:   "github",
-				Type:     "pull_request_assigned",
-				Assignee: "octocat",
-			},
-			want: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.DeepEqual(t, tt.event.MatchRule(tt.rule), tt.want)
-		})
-	}
-}
 
 func TestInputEventAttr(t *testing.T) {
 	e := InputEvent{
